requirements.txt
requests
flask
flask-bcrypt 
flask-sqlalchemy 
python-dotenv
mailjet-rest